moving the swagger api generation to regular test suite
