+++ api/src/test/java/uk/gov/hmcts/payment/api/componenttests/SwaggerPublisher.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.payment.integration;
+package uk.gov.hmcts.payment.api.componenttests;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -8,6 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.payment.api.configuration.security.AuthCheckerConfiguration;
 
@@ -15,14 +16,14 @@ import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
+import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.MOCK;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-
-@RunWith(SpringJUnit4ClassRunner.class)
+@RunWith(SpringRunner.class)
+@ActiveProfiles({"local", "componenttest", "mockcallbackservice"})
+@SpringBootTest(webEnvironment = MOCK)
 @AutoConfigureMockMvc
-@SpringBootTest//(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-//@ContextConfiguration(classes = IntegrationTestContextConfiguration.class)
 public class SwaggerPublisher {
 
     @Autowired
